+++ interfaces/native/vibrator/src/vibrator_agent.cpp
@@ -92,7 +92,7 @@ int32_t PlayVibratorCustom(int32_t fd, int64_t offset, int64_t length)
 #ifdef OHOS_BUILD_ENABLE_VIBRATOR_CUSTOM
     if (fd < 0 || offset < 0 || length <= 0) {
         MISC_HILOGE("input parameter invalid, fd:%{public}d, offset:%{public}lld, length:%{public}lld",
-            fd, offset, length);
+            fd, static_cast<long long>(offset), static_cast<long long>(length));
         return PARAMETER_ERROR;
     }
     auto &client = VibratorServiceClient::GetInstance();
